third_party/WebKit/LayoutTests/fast/dom/idl-dictionary-unittest-expected.txt

IDL dictionary unittest

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test for setting an empty dictionary
PASS dict.longMember is undefined.
PASS dict.longMemberWithDefault is 42
PASS dict.longOrNullMember is undefined.
PASS dict.longOrNullMemberWithDefault is null
PASS dict.booleanMember is undefined.
PASS dict.doubleMember is undefined.
PASS dict.stringMember is undefined.
PASS dict.stringMemberWithDefault is "defaultStringValue"
PASS dict.stringSequenceMember is undefined.
PASS dict.stringSequenceOrNullMember is undefined.
PASS dict.elementMember is undefined.
PASS dict.elementOrNullMember is undefined.
PASS dict.enumMember is undefined.
PASS dict.enumMemberWithDefault is "foo"
PASS dict.enumOrNullMember is undefined.

Test for setting undefined
PASS dict.longMember is undefined.
PASS dict.longMemberWithDefault is 42
PASS dict.longOrNullMember is undefined.
PASS dict.longOrNullMemberWithDefault is null
PASS dict.booleanMember is undefined.
PASS dict.doubleMember is undefined.
PASS dict.stringMember is undefined.
PASS dict.stringMemberWithDefault is "defaultStringValue"
PASS dict.stringSequenceMember is undefined.
PASS dict.stringSequenceOrNullMember is undefined.
PASS dict.elementMember is undefined.
PASS dict.elementOrNullMember is undefined.

Test for setting valid values
PASS dict.longMember is 1
PASS dict.longMemberWithDefault is 2
PASS dict.longOrNullMember is 3
PASS dict.longOrNullMemberWithDefault is 4
PASS dict.stringMember is "modifiedString1"
PASS dict.stringMemberWithDefault is "modifiedString2"
PASS dict.booleanMember is true
PASS dict.doubleMember is 3.14
PASS dict.stringSequenceMember is ["foo", "bar", "baz"]
PASS dict.stringSequenceOrNullMember is []
PASS dict.elementMember is element1
PASS dict.elementOrNullMember is undefined.
PASS dict.enumMember is "foo"
PASS dict.enumMemberWithDefault is "bar"
PASS dict.enumOrNullMember is "baz"

Test for explicit undefined or null, and missing members
PASS dict.longMember is undefined.
PASS dict.longMemberWithDefault is 42
PASS dict.longOrNullMember is undefined.
PASS dict.longOrNullMemberWithDefault is null

Test for setting invalid member
PASS dict.invalidMember is undefined.

Test for setting invalid enum value
PASS dictionaryTest.set({enumMember: 'invalid'}) threw exception TypeError: Failed to execute 'set' on 'DictionaryTest': member enumMember ('invalid') is not a valid enum value..

Test for passing invalid dictionary values
PASS dictionaryTest.set(42) threw exception TypeError: Failed to execute 'set' on 'DictionaryTest': parameter 1 ('testingDictionary') is not an object..
PASS dictionaryTest.set('string') threw exception TypeError: Failed to execute 'set' on 'DictionaryTest': parameter 1 ('testingDictionary') is not an object..

PASS successfullyParsed is true

TEST COMPLETE
